Adding the most basic structure of the project
Create your component inside of the "components" folder, and if you function requires you to create a class, please do so by creating a folder inside of src called "classes", I will be adding more stuff to the raw structure as the days go by, but if any major change is pushed you will be let known. Thank you!